feature: Fixed the positioning of enemies
If you notice, each sprite has a horizontal and vertical center. Since
we are dealing with a discrete horizontal spaces, it is critical to
align the characters with respect to the horizontal spaces, which are
the blocks. The center of a character just needs to be positioned at a
block's center. This can be done by displacing the character's center
to match with the block's center.