bug: fix collision animation
Previously, the animation occurred too early. This is because there is
blank spaces on the images and the collision would think that the body
of the sprite includes the blank spaces. For the animations, I took
account of the white spaces. I did this by adding a new class called
CharacterSprite, which takes into account the position of the left and
the right side of a character.